UNG-1865 - * Little bit of reorganizing code * Further updates to README
