--- src/main/scala/com/ubirch/db/models/User.scala
@@ -1,6 +1,0 @@
-package com.ubirch.db.models
-import com.ubirch.models.user.Email
-
-import java.util.UUID
-
-case class User(id: UUID, email: Email, status: UserStatus = WaitingForRequiredActions)
