+++ src/main/scala/com/ubirch/models/keycloak/user/KeycloakUser.scala
@@ -1,6 +1,6 @@
 package com.ubirch.models.keycloak.user
 
-import com.ubirch.models.user.{Email, FirstName, LastName}
+import com.ubirch.models.user.{Email, FirstName, LastName, UserId, UserName}
 
 case class KeycloakUser(
   id: UserId,
