+++ src/main/scala/com/ubirch/models/keycloak/user/UserException.scala
@@ -1,4 +1,5 @@
 package com.ubirch.models.keycloak.user
+import com.ubirch.models.user.UserName
 
 sealed trait UserException
 case class UserAlreadyExists(uerName: UserName) extends UserException
