--- src/main/scala/com/ubirch/models/keycloak/user/UserId.scala
@@ -1,5 +1,0 @@
-package com.ubirch.models.keycloak.user
-
-import java.util.UUID
-
-case class UserId(value: UUID) extends AnyVal
