+++ src/main/scala/com/ubirch/models/user/User.scala
@@ -1,0 +1,3 @@
+package com.ubirch.models.user
+
+case class User(id: UserId, email: Email, status: UserStatus = WaitingForRequiredActions)
