+++ src/main/scala/com/ubirch/models/user/UserStatus.scala
@@ -1,4 +1,5 @@
-package com.ubirch.db.models
+package com.ubirch.models.user
+
 import io.getquill.MappedEncoding
 
 sealed trait UserStatus
