+++ src/test/scala/com/ubirch/data/KeycloakTestData.scala
@@ -2,8 +2,8 @@ package com.ubirch.data
 
 import com.ubirch.models.keycloak.group.{CreateKeycloakGroup, GroupName}
 import com.ubirch.models.keycloak.roles.{CreateKeycloakRole, RoleName}
-import com.ubirch.models.keycloak.user.{CreateKeycloakUser, UserName}
-import com.ubirch.models.user.{Email, FirstName, LastName}
+import com.ubirch.models.keycloak.user.CreateKeycloakUser
+import com.ubirch.models.user.{Email, FirstName, LastName, UserName}
 
 import scala.util.Random
