+++ src/test/scala/com/ubirch/db/DBIntegrationTest.scala
@@ -1,9 +1,8 @@
 package com.ubirch.db
 
 import com.ubirch.E2ETestBase
-import com.ubirch.db.models.{User, WaitingForRequiredActions}
 import com.ubirch.db.tables.UserRepository
-import com.ubirch.models.user.Email
+import com.ubirch.models.user.{Email, User, UserId, WaitingForRequiredActions}
 
 import java.util.UUID
 import scala.concurrent.duration.DurationInt
@@ -14,7 +13,7 @@ class DBIntegrationTest extends E2ETestBase {
     "should be able to store and retrieve data in DB" in {
       withInjector { injector =>
         val repo = injector.get[UserRepository]
-        val userId = UUID.randomUUID()
+        val userId = UserId(UUID.randomUUID())
         val userEmail = Email("testmail@example.com")
         val res = for {
           _ <- repo.createUser(User(userId, userEmail))
